RTC implementation in irrigation system.
This module will integrate the RTC, Hardware in the Irrigation system
the RTC module PN is. DS1307 and will used as second clock option.

RTC implementation in irrigation system.

This module will integrate the RTC, Hardware in the Irrigation system
the RTC module PN is. DS1307 and will used as second clock option.

Signed-off-by: Preciado <jose.preciado@intel.com>